+++ pkg/oauth/jwt_manager.go
@@ -2,7 +2,6 @@ package oauth
 
 import (
 	"github.com/hellofresh/janus/pkg/jwt"
-	"github.com/hellofresh/janus/pkg/session"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -16,31 +15,13 @@ func NewJWTManager(parser *jwt.Parser) *JWTManager {
 	return &JWTManager{parser}
 }
 
-// Set returns nil since when we work with JWT we don't need to store them
-func (m *JWTManager) Set(accessToken string, session session.State, resetTTLTo int64) error {
-	return nil
-}
-
-// Remove returns nil because there is not storage to remove from
-func (m *JWTManager) Remove(accessToken string) error {
-	return nil
-}
-
-// IsKeyAuthorised checks if the access token is valid
-func (m *JWTManager) IsKeyAuthorised(accessToken string) (session.State, bool) {
-	var sessionState session.State
-
-	token, err := m.parser.Parse(accessToken)
+// IsKeyAuthorized checks if the access token is valid
+func (m *JWTManager) IsKeyAuthorized(accessToken string) bool {
+	_, err := m.parser.Parse(accessToken)
 	if err != nil {
 		log.WithError(err).Info("Failed to parse and validate the JWT")
-		return sessionState, false
+		return false
 	}
 
-	// as parser.Parse() does validation we are sure that token is valid at this point
-	if claims, ok := m.parser.GetMapClaims(token); ok {
-		sessionState.AccessToken = accessToken
-		sessionState.ExpiresIn = claims["exp"].(int64)
-	}
-
-	return sessionState, true
+	return true
 }