+++ pkg/oauth/oauth.go
@@ -31,23 +31,37 @@ type OAuth struct {
 	AllowedAuthorizeTypes  []AuthorizeRequestType `bson:"allowed_authorize_types" json:"allowed_authorize_types" mapstructure:"allowed_authorize_types"`
 	AuthorizeLoginRedirect string                 `bson:"auth_login_redirect" json:"auth_login_redirect" mapstructure:"auth_login_redirect"`
 	Secrets                map[string]string      `bson:"secrets" json:"secrets"`
-	CorsMeta               meta                   `bson:"cors_meta" json:"cors_meta" valid:"cors_meta" mapstructure:"cors_meta"`
+	CorsMeta               corsMeta               `bson:"cors_meta" json:"cors_meta" mapstructure:"cors_meta"`
+	RateLimit              rateLimitMeta          `bson:"rate_limit" json:"rate_limit"`
 	TokenStrategy          TokenStrategy          `bson:"token_strategy" json:"token_strategy" mapstructure:"token_strategy"`
 	AccessRules            []*AccessRule          `bson:"access_rules" json:"access_rules"`
 }
 
 // Endpoints defines the oauth endpoints that wil be proxied
 type Endpoints struct {
-	Authorize *proxy.Definition `bson:"authorize" json:"authorize"`
-	Token     *proxy.Definition `bson:"token" json:"token"`
-	Info      *proxy.Definition `bson:"info" json:"info"`
-	Revoke    *proxy.Definition `bson:"revoke" json:"revoke"`
+	Authorize  *proxy.Definition `bson:"authorize" json:"authorize"`
+	Token      *proxy.Definition `bson:"token" json:"token"`
+	Introspect *proxy.Definition `bson:"introspect" json:"introspect"`
+	Revoke     *proxy.Definition `bson:"revoke" json:"revoke"`
 }
 
 // ClientEndpoints defines the oauth client endpoints that wil be proxied
 type ClientEndpoints struct {
 	Create *proxy.Definition `bson:"create" json:"create"`
 	Remove *proxy.Definition `bson:"remove" json:"remove"`
+}
+
+type rateLimitMeta struct {
+	Limit   string `bson:"limit" json:"limit"`
+	Enabled bool   `bson:"enabled" json:"enabled"`
+}
+
+type corsMeta struct {
+	Domains        []string `mapstructure:"domains" bson:"domains" json:"domains"`
+	Methods        []string `mapstructure:"methods" bson:"methods" json:"methods"`
+	RequestHeaders []string `mapstructure:"request_headers" bson:"request_headers" json:"request_headers"`
+	ExposedHeaders []string `mapstructure:"exposed_headers" bson:"exposed_headers" json:"exposed_headers"`
+	Enabled        bool     `bson:"enabled" json:"enabled"`
 }
 
 // TokenStrategy defines the token strategy fields
@@ -76,14 +90,6 @@ func (t TokenStrategy) GetJWTSigningMethods() ([]jwt.SigningMethod, error) {
 		return []jwt.SigningMethod{{Alg: "HS256", Key: legacy.Secret}}, nil
 	}
 	return methods, err
-}
-
-type meta struct {
-	Domains        []string `mapstructure:"domains" bson:"domains" json:"domains"`
-	Methods        []string `mapstructure:"methods" bson:"methods" json:"methods"`
-	RequestHeaders []string `mapstructure:"request_headers" bson:"request_headers" json:"request_headers"`
-	ExposedHeaders []string `mapstructure:"exposed_headers" bson:"exposed_headers" json:"exposed_headers"`
-	Enabled        bool     `bson:"enabled" json:"enabled"`
 }
 
 // AccessRule represents a rule that will be applied to a JWT that could be revoked